Modify New Book form and Book constructor function
Revised final section of 'New Book' form in HTML to reflect whether or
not the user completed the book. The 'Phone Number' option was simply
a placeholder to get the form going prior. Added styling comments to
CSS in an attempt to center the form at the bottom of the page under
the 'New Book' button, but this broke button functionality in the form.
Updated the 'Book' constructor function with 'No' prompts and return
statements to line up with the other functions in the JS that were
looking for 'No' as a user input. Added further comments in other
existing functions for future reference. Deleted comments that were no
longer needed.