Modify HTML and JS for debugging errors
Modified the button type on the Confirm button in order to stop a
'not focusable' form error that was popping up in debugging.
Attempted to debug Confirm button and addBookToLibrary function
several times with different debug attempts in the main code. Confirm
button still not functioning correctly and don't know why.