Modify functions to resolve errors while debugging
"Finished" status was not posting correctly on the addBookToLibrary
function. Modified HTML to resolve this error. Spread syntax on the
push function turned out to add another error, so it was removed to
resolve this. The function seems to run properly now. "library" was
not defined in the DOM for the bookList function. Another value needs
to be chose. So far, "table" inside the function is the placeholder,
but will experiment with other DOM targets in the future. Still an
issue with firstEmptyRow as well - entire for loop gets skipped and
function runs straight to the bottom where an attribute for the first
empty row tries to be set. Added more comments as notes/future
instructions throughout for codebase revisit.