Modify bookList and confirm button functions
Made several modifications on the bookList function in an attempt to
fix the error of the user input not generating onto the table on the
webpage. Tried to target different elements & use a new conditional
statement for populating the text. Error still persists but have other
methods to try in the comments. Moved dialog code out of bookList into
the confirm button function as that seems to make more sense.