Modify bookList function further
Continued to iterate and try different arguments to get the for loop
to work on the bookList function so the user info would post to the
table row on the webpage. None of the targets involving the table
elements work. Commented in the function with next steps to try.
Also moved commented code out of button functions to their proper places
per previous comments.