Modify bookDisplay & fix errors in function
Was able to finally get bookDisplay to work somewhat correctly. The
layout is still off, but it seems to be more of an HTML problem at
this point than a JS issue. All user input posts to the table
correctly EXCEPT for Finished status (shows actual function code).
Added comments with next steps such as Remove button repositioning &
layout in bookDisplay etc.