Fix remove button errors in bookDisplay function
Remove button was not being dynamically added properly due to
incorrect targeting on the element. Existing element was being
targeted instead of a new button element being created. Modified all
REMOVE button functionality to reflect this and reordered DOM
manipulation elements in bookDisplay so that information would generate
in the correct order on user submit.