Modify Book constructor & comment next steps
Removed function for this.read in the Book constructor which resolved
the "read: function" issue when read status was posted to the table
through user data. Commented issues and possible next steps for
table row positioning of new books added, remove button functionality &
error of looping through entire array instead of just latest object
when book data is pushed to the table.