Modify remove button function with temporary fix
Fixed incorrect TypeError with remove() on the clearRow function as a
temporary fix for the Remove buttons on the table rows. Remove buttons
work but ALL table rows are removed instead of the row specifice to the
Remove button that was clicked. Commented possible solution to the
bookDisplay function looping through the entire array every time user
data is entered. Comments are inside the bookDisplay itself (forEach
method)