Modify Confirm button functionality
Added closing functionality to the Confirm button so the dialog now
closes once all user data entered is valid and placed into the table.
Attempted to solve the issue of the entire array loop with
replaceChildren function in addBookToLibrary, but replaceChildren is
not a true solution. It deletes any row already entered only for
bookDisplay to bring it back along with the next array object when the
loop finishes (same end result as before).