Debug & add comments for table row positioning
Made several attempts targeting table row to get the generated rows
of info to appear as a column unsuccessfully. Commented notes related
to dev comments in HTML on next steps to solve this issue. Added
Remove button styling from JS back into external CSS as comments to
debug why styling didn't work when it was external but did when it was
in JS bookDisplay function.