Comment next steps & possible solutions
Added dev comments with more detail on the positioning issue in the
HTML and reminder to possibly delete what's been tried already in
commented blocks. Added dev comments in CSS on possible solution to
form dialog not being centered on the page. Added dev comments in JS
with updates to positioning issue in more detail, possible solutions in
bookDisplay and possible code addition in clearRow function if needed.