Modify and resolve remove button functionality
Made several more attempts to fix the issue of the remove button erasing
the entire table instead of just the related table row the button was
on. Revisited some code and made one small modification which causes
all buttons to work properly. Updated and deleted dev comments related
to remove button issue as they are either no longer needed or may be
useful in the future.