Modify table row styling for loop
Made several attempts to change the color of every "even" table row in
the app to be slightly darker than the standard gray we have right now.
Was unsuccessful after trying bitwise, Math.floor and even a different
variation of modulus method (which is left in there for now since it
doesn't affect the rest of the code). Also tried to add multiple array
items as arguments for a previous solution but didn't work either as
it throws an error.