Fix Remove & Read button styling issue
Was finally able to figure out why none of the separate CSS styling
attempts I made in the past were working on the Remove buttons. This
was due to dot notation I had on the addClassList in the JS that I
didn't need at all (the separate styling worked on the Read button
because it didn't have the dot notation). Removed all styling from the
JS and placed back in the CSS doc for the Remove and Read (Finished?)
buttons.