Modify "Finished?" status buttons
Was able to get the functionality for each "Finished?" button working.
They now change the status of "Finished" from "Yes" to "No" or vice
versa if the user clicks the button. The issue now is that the buttons
only work once. The user can't click the button again to switch back to
the previous "Finished" status, for example. Will continue to debug.